requirements.txt
python-gitlab>=4.0.0
PyYAML>=6.0
rich>=13.0.0
prompt-toolkit>=3.0.0
typer>=0.9.0 